Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pydantic==2.9.2
python-dotenv==1.0.1
